feat: add more useful just recipes
