+++ .env
@@ -1,1 +1,2 @@
 SOROBAN_CONTRACT_ID=a0ac4c26215ee63cbf9084ae2aa369df3808cf4a2a3f57a1da07956562791258
+CONFIG_DIR=.
